Only show current_users todos.

+++ flask_app.py
@@ -114,20 +114,20 @@ def logout():
 def index():
     # GET
     if request.method == "GET":
-        todos = db_read("SELECT id, content, due FROM todos ORDER BY due")
+        todos = db_read("SELECT id, content, due FROM todos WHERE user_id=%s ORDER BY due", (current_user.id,))
         return render_template("main_page.html", todos=todos)
 
     # POST
     content = request.form["contents"]
     due = request.form["due_at"]
-    db_write("INSERT INTO todos (content, due) VALUES (%s, %s)", (content, due, ))
+    db_write("INSERT INTO todos (user_id, content, due) VALUES (%s, %s)", (current_user.id, content, due, ))
     return redirect(url_for("index"))
 
 @app.post("/complete")
 @login_required
 def complete():
     todo_id = request.form.get("id")
-    db_write("DELETE FROM todos WHERE id=%s", (todo_id,))
+    db_write("DELETE FROM todos WHERE user_id=%s AND id=%s", (current_user.id, todo_id,))
     return redirect(url_for("index"))
 
 if __name__ == "__main__":
